Backend API

+++ server/models/order.go
@@ -7,7 +7,7 @@ import (
 type Product struct {
 	Prod_ID primitive.ObjectID `json:"_id,omitempty" bson:"_id,omitempty"`
 	Name    string             `json:"name,omitempty" bson:"name,omitempty"`
-	Price   int32              `json:"price,omitempty" bson:"price,omitempty"`
+	Price   float64              `json:"price,omitempty" bson:"price,omitempty"`
 }
 
 type Customer struct {
